MTSDK 309 - Extract log messages to separate file
-Convert all log.w
-Correct some LogMessages.kt function parameters
-Correct most of logs in tests classes
-Move LogTag to TestValues.kt
-Comment out previous LogMessages for tests (for now)

+++ transport/src/iosMain/kotlin/com/genesys/cloud/messenger/transport/network/PlatformSocket.kt
@@ -72,7 +72,7 @@ internal actual class PlatformSocket actual constructor(
                     reason: NSData?,
                 ) {
                     val why = reason?.string() ?: "Reason not specified."
-                    log.i { LogMessages.socketDidClose(didCloseWithCode.toInt(), why, active) }
+                    log.i { LogMessages.socketDidClose(didCloseWithCode, why, active) }
                     if (webSocketTask == webSocket) {
                         deactivate()
                         listener.onClosed(code = didCloseWithCode.toInt(), reason = why)
